requirements.txt
# QA Auto Test Builder - Dependencies
# 버전 호환성을 위해 최소/최대 버전 지정

# 브라우저 자동화
selenium>=4.0.0,<5.0.0
webdriver-manager>=4.0.0,<5.0.0

# 테스트 프레임워크
pytest>=7.0.0,<9.0.0
pytest-xdist>=3.0.0,<4.0.0

# 리포팅
allure-pytest>=2.13.0,<3.0.0

# 데이터 처리
pandas>=1.5.0,<3.0.0
openpyxl>=3.0.0,<4.0.0

# 보안
cryptography>=41.0.0,<43.0.0
